[#124] 리팩토링
- 패키지명 수정
- 변수 추출
- 불필요한 변수 제거

+++ src/main/java/io/driver/codrive/modules/room/service/ImageService.java
@@ -28,7 +28,7 @@ import lombok.extern.slf4j.Slf4j;
 public class ImageService {
 	private static final List<String> ALLOWED_EXTENSIONS = List.of("jpg", "jpeg", "png", "gif", "webp", "bmp");
 	private static final String DIR_PATH = "images/";
-	private static final long MAX_FILE_SIZE = 5 * 1024 * 1024;
+	private static final int MAX_FILE_SIZE = 5 * 1024 * 1024;
 	private final AmazonS3 amazonS3Client;
 
 	@Value("${cloud.aws.s3.bucket}")
